refactor: キャンプ場登録と更新は同じユースケースで対応できるのでupdateだけにした
