+++ app/UseCase/CampGrounds/UpdateCampGroundCommand.php
@@ -19,4 +19,34 @@ class UpdateCampGroundCommand
         $this->image = $image;
         $this->status = $status;
     }
+
+    public function getId(): int
+    {
+        return $this->id;
+    }
+
+    public function getName(): string
+    {
+        return $this->name;
+    }
+
+    public function getAddress(): string
+    {
+        return $this->address;
+    }
+
+    public function getPrice(): int
+    {
+        return $this->price;
+    }
+
+    public function getImage(): string
+    {
+        return $this->image;
+    }
+
+    public function getStatus(): string
+    {
+        return $this->status;
+    }
 }
